refactor(CA-771): move DateRange entity out of the DateRangeBottomSheet widget file into search_filters domain
