+++ test/libraries/search_filters/widgets/date_filter_chip_test.dart
@@ -1,6 +1,6 @@
 import 'package:construculator/l10n/generated/app_localizations.dart';
+import 'package:construculator/libraries/search_filters/domain/entities/date_range.dart';
 import 'package:construculator/libraries/search_filters/presentation/widgets/date_filter_chip.dart';
-import 'package:construculator/libraries/search_filters/presentation/widgets/date_range_bottom_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
